requirements.txt
azure-mgmt-resource
azure-identity
azure-mgmt-automanage